Move CoreVisitor from core.py into parse.py (its sibling)
CoreVisitor (the parsimonious NodeVisitor that turns parsed bigraph
expressions into schema nodes, ~114 lines) lived in core.py but is a parsing
backend — it belongs next to parse.py's other visitors (ParameterVisitor,
visit_expression). It has no external importers (only constructed inside Core
at parse_visitor = CoreVisitor(self) / CoreVisitor(clone)).

Moved to parse.py with its deps (Union/Tuple/Node from schema, handle_parameters
from methods); core.py imports CoreVisitor back from parse and drops the now-
unused NodeVisitor + handle_parameters imports. Verified no import cycle (neither
schema nor methods imports parse) and that core.access(...) — the grammar path
that uses CoreVisitor — still works.

Pure structural move, no behavior change. Full suite: 1040 passed.
